+++ CHANGELOG.md
@@ -12,6 +12,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 
+- Fixed interrupt token handling in `VoiceOverView` that would cause it to permanently stop a Dialogue Runner's ability to progress through dialogue lines.
+
 - Space sample now has two scenes to show both approaches to localisation.
   - `Space-UnityLocalization.unity` uses the Unity Localization package
   - `Space-InbuiltLocalisation.unity` uses the inbuilt localisation system.